requirements.txt
flask
flask-cors
flask-socketio
flask-bcrypt
flask-jwt-extended
Flask-SQLAlchemy

hl7apy>=1.3.0
google-generativeai
google-auth
python-dotenv

eventlet
gunicorn

sqlalchemy
psycopg2-binary # PostgreSQL Driver
alembic  

pydantic
pydantic-settings

requests